requirements.txt
Django>=4.2.0
django-allauth>=0.54.0
Pillow>=8.0.0
python-decouple>=3.6
requests>=2.25.0
qrcode[pil]>=7.3.1
whitenoise>=6.0.0

# Video generation dependencies
replicate>=0.15.0
moviepy>=1.0.3
celery[redis]>=5.3.0
redis>=4.5.0

# Optional: For better video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Development dependencies (optional)
django-debug-toolbar>=4.0.0
